Fixed lookup of XLIFF files
pac:xliff() now returns an absolute URI resolved against the xml:base
of the element where the XLIFF file is referenced. This is so Saxon
can resolve the absolute URI of the XLIFF file against an XML catalog.

Added XSLT Standard Library for the uri:resolve-uri template.

Fixed XlateConvert.xsl and XlateCombine.xsl to they no longer attempt
to resolve location of XLIFF file against the root of the source doc.